Create Policies for all models

+++ app/Http/Controllers/AnswerController.php
@@ -6,6 +6,7 @@ use App\Answer;
 use App\Block;
 use App\Question;
 use App\Survey;
+use App\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 
@@ -26,30 +27,47 @@ class AnswerController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
+     * @param Answer $answer
+     * @param  \Illuminate\Http\Request $request
+     * @param User $user
      * @return \Illuminate\Http\Response
+     * @internal param int $id
      */
-    public function update(Answer $answer, Request $request)
+    public function update(Answer $answer, Request $request, User $user)
     {
-        if ($answer->question->hasRadioAnswer()){
-            $answer->update($request->all());
+        if ($user->can('update', $answer)) {
+            if ($answer->question->hasRadioAnswer()){
+                $answer->update($request->all());
+            }
+            return compact('answer');
+        }else{
+            abort(404);
         }
-        return compact('answer');
-
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param Survey $survey
+     * @param Block $block
+     * @param Question $question
+     * @param Answer $answer
+     * @param User $user
      * @return \Illuminate\Http\Response
+     * @internal param int $id
      */
-    public function destroy(Survey $survey, Block $block, Question $question, Answer $answer)
+    public function destroy(Survey $survey, Block $block, Question $question, Answer $answer, User $user)
     {
-        if ($answer->question->hasRadioAnswer()) {
-            $answer->delete();
+        if ($user->can('delete', $answer)) {
+            if ($answer->question->hasRadioAnswer()) {
+                $answer->delete();
+            }
+            return compact('answer');
+        }else{
+            abort(404);
         }
-        return compact('answer');
+
+
+
     }
 }
