+++ app/Http/Controllers/BlockController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Block;
 use App\Customer;
 use App\Survey;
+use App\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 
@@ -39,25 +40,36 @@ class BlockController extends Controller
      *
      * @param Block $block
      * @param Request $request
+     * @param User $user
      * @return array
      */
-    public function update(Block $block, Request $request)
+    public function update(Block $block, Request $request, User $user)
     {
-        $block->update($request->all());
+        if ($user->can('update', $block)) {
+            $block->update($request->all());
 
-        return compact('block');
+            return compact('block');
+        }else{
+            abort(404);
+        }
+
     }
 
     /**
      * Remove the specified resource from storage.
      *
      * @param Block $block
+     * @param User $user
      * @return array
      */
-    public function destroy(Block $block)
+    public function destroy(Block $block, User $user)
     {
-        $block->delete();
-        return compact('block');
+        if ($user->can('delete', $block)) {
+            $block->delete();
+            return compact('block');
+        }else{
+            abort(404);
+        }
     }
