+++ app/Http/Controllers/QuestionController.php
@@ -6,6 +6,7 @@ use App\Block;
 use App\Customer;
 use App\Question;
 use App\Survey;
+use App\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 
@@ -28,45 +29,60 @@ class QuestionController extends Controller
      *
      * @param Question $question
      * @param Request $request
+     * @param User $user
      * @return array
      */
-    public function addAnswer(Question $question, Request $request)
+    public function addAnswer(Question $question, Request $request, User $user)
     {
-        if($question->hasRadioAnswer()) {
-            $answer = $question->answer()->create($request->all());
+        if ($user->can('addAnswer', $question)) {
+            if($question->hasRadioAnswer()) {
+                $answer = $question->answer()->create($request->all());
 
-            return compact('answer');
-            //TODO it is return true 200
+                return compact('answer');
+                //TODO it is return true 200
+            }
+            else {
+                return false;
+            }
+        }else{
+            abort(404);
+        }
 
-        }
-        else {
-            return false;
-        }
     }
 
     /**
      * Update question
      *
-     * @param Request $request
      * @param Question $question
+     * @param Request $request
+     * @param User $user
      * @return array
      */
-    public function update(Question $question, Request $request)
+    public function update(Question $question, Request $request, User $user)
     {
-        $question->update($request->all());
-        return compact('question');
+        if ($user->can('update', $question)) {
+            $question->update($request->all());
+            return compact('question');
+        }else{
+            abort(404);
+        }
     }
 
     /**
      * Remove the specified resource from storage.
      *
      * @param Question $question
+     * @param User $user
      * @return array
      */
-    public function destroy(Question $question)
+    public function destroy(Question $question, User $user)
     {
-        $question->delete();
-        return compact('question');
+        if ($user->can('delete', $question)) {
+            $question->delete();
+            return compact('question');
+        }else{
+            abort(404);
+        }
     }
